Подправил js Реформат кода Проверка на админа

+++ app/controllers/StudentsController.php
@@ -21,6 +21,9 @@ class StudentsController extends BaseController
 
     public function create()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $data['title'] = 'Создание студента';
         $data['content'] = 'students/create';
         $data['data'] = (new Groups())->findAll();
@@ -30,16 +33,18 @@ class StudentsController extends BaseController
 
     public function store()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $data['name'] = $this->request->getPost('name');
         $data['surname'] = $this->request->getPost('surname');
-        $data['group_id'] =$this->request->getPost('group_id');
+        $data['group_id'] = $this->request->getPost('group_id');
         $obj = new Students();
         try {
             $obj->insert($data);
-        } catch (DatabaseException $e)
-        {
+        } catch (DatabaseException $e) {
             $this->session->set([
-                'msg' => 'Ошибка!'.$e->getMessage(),
+                'msg' => 'Ошибка!' . $e->getMessage(),
                 'msg_type' => 'alert-danger'
             ]);
             return $this->response->redirect(site_url('/students'));
@@ -49,10 +54,12 @@ class StudentsController extends BaseController
 
     public function edit(int $id)
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $obj = new Students();
         $data['data'] = $obj->find($id);
-        if ($data['data'] == null)
-        {
+        if ($data['data'] == null) {
             $this->session->set([
                 'msg' => 'Ошибка! Студент не найден',
                 'msg_type' => 'alert-danger'
@@ -68,6 +75,9 @@ class StudentsController extends BaseController
 
     public function update()
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $id = $this->request->getPost('id');
         $data['name'] = $this->request->getPost('name');
         $data['surname'] = $this->request->getPost('surname');
@@ -75,10 +85,9 @@ class StudentsController extends BaseController
         $obj = new Students();
         try {
             $obj->update($id, $data);
-        } catch (DatabaseException $e)
-        {
+        } catch (DatabaseException $e) {
             $this->session->set([
-                'msg' => 'Ошибка!'.$e->getMessage(),
+                'msg' => 'Ошибка!' . $e->getMessage(),
                 'msg_type' => 'alert-danger'
             ]);
             return $this->response->redirect(site_url('/students'));
@@ -88,8 +97,11 @@ class StudentsController extends BaseController
 
     public function delete(int $id)
     {
+        if (!$this->session->get('isAdmin')) {
+            return $this->response->redirect('/students');
+        }
         $obj = new Students();
-        $obj->delete($id); # INFO Интересно, что бд не вызывает исключение если удалять несуществующий id
+        $obj->delete($id);
         return $this->response->redirect(site_url('/students'));
     }
 }
